feat: [factory-argument-signatures-prd-009] - [Render signature-backed invocation forms in the dashboard]
