+++ pkg/factory/validation/validate.go
@@ -2,8 +2,6 @@ package validation
 
 import (
 	"fmt"
-	"regexp"
-	"slices"
 	"strings"
 
 	factoryapi "github.com/portpowered/infinite-you/pkg/api/generated"
@@ -147,613 +145,6 @@ func invocationReturnTarget(code, field, message string) Target {
 			Location: SubjectLocationDefinition,
 		},
 		Path: fmt.Sprintf("%s.invocationReturn.%s", validationRoot, field),
-	}
-}
-
-var invocationSignatureReferencePattern = regexp.MustCompile(`\$\{([A-Za-z0-9_.-]+)\}`)
-
-var supportedInvocationTypeHints = []string{
-	string(factoryapi.FactoryInvocationParameterTypeHintString),
-	string(factoryapi.FactoryInvocationParameterTypeHintPath),
-	string(factoryapi.FactoryInvocationParameterTypeHintFilePath),
-	string(factoryapi.FactoryInvocationParameterTypeHintDirectoryPath),
-	string(factoryapi.FactoryInvocationParameterTypeHintNumberString),
-	string(factoryapi.FactoryInvocationParameterTypeHintBooleanString),
-}
-
-var supportedInvocationValueModes = []string{
-	string(factoryapi.FactoryInvocationParameterValueModeExact),
-	string(factoryapi.FactoryInvocationParameterValueModeRepeated),
-	string(factoryapi.FactoryInvocationParameterValueModeVariadic),
-	string(factoryapi.FactoryInvocationParameterValueModeFileContents),
-}
-
-var supportedInvocationBindingKinds = []string{
-	string(factoryapi.FactoryInvocationParameterBindingKindPositional),
-	string(factoryapi.FactoryInvocationParameterBindingKindNamed),
-	string(factoryapi.FactoryInvocationParameterBindingKindStdin),
-	string(factoryapi.FactoryInvocationParameterBindingKindNamedRest),
-}
-
-var supportedUnknownNamedArgumentPolicies = []string{
-	string(factoryapi.FactoryInvocationUnknownNamedArgumentPolicyReject),
-	string(factoryapi.FactoryInvocationUnknownNamedArgumentPolicyAllow),
-	string(factoryapi.FactoryInvocationUnknownNamedArgumentPolicyCollect),
-}
-
-var supportedOutputContractModes = []string{
-	string(factoryapi.FactoryInvocationOutputContractModeInline),
-	string(factoryapi.FactoryInvocationOutputContractModeFile),
-	string(factoryapi.FactoryInvocationOutputContractModeJson),
-}
-
-type invocationParameterSummary struct {
-	name      string
-	valueMode string
-}
-
-type interpolationFieldTarget struct {
-	subjectType     SubjectType
-	subjectID       string
-	path            string
-	location        SubjectLocation
-	value           string
-	allowsRepeated  bool
-	fieldDescriptor string
-}
-
-// InvocationSignatureTargets validates the public invocation signature contract
-// and supported ${parameter} references before runtime normalization runs.
-func InvocationSignatureTargets(cfg *interfaces.FactoryConfig) []Target {
-	if cfg == nil || cfg.InvocationSignature == nil {
-		return nil
-	}
-
-	signature := cfg.InvocationSignature
-	var targets []Target
-
-	policy := strings.TrimSpace(signature.UnknownNamedArgumentPolicy)
-	if policy != "" && !slices.Contains(supportedUnknownNamedArgumentPolicies, policy) {
-		targets = append(targets, invocationSignatureTarget(
-			CodeInvocationSignatureUnsupportedUnknownNamedArgumentPolicy,
-			"unknownNamedArgumentPolicy",
-			fmt.Sprintf("invocationSignature.unknownNamedArgumentPolicy %q is not supported", signature.UnknownNamedArgumentPolicy),
-		))
-	}
-
-	parametersByName := make(map[string]invocationParameterSummary, len(signature.Parameters))
-	namedKeys := map[string]string{}
-	positionalNames := map[int]string{}
-	var positionalSlots []int
-	stdinParameters := []string{}
-	namedRestParameters := []string{}
-	var variadicPositionalNames []string
-
-	for index, parameter := range signature.Parameters {
-		targets = append(targets, invocationSignatureParameterTargets(parameter, index)...)
-
-		name := strings.TrimSpace(parameter.Name)
-		if _, exists := parametersByName[name]; exists {
-			targets = append(targets, invocationSignatureParameterTarget(
-				CodeInvocationSignatureDuplicateParameterName,
-				index,
-				"name",
-				name,
-				fmt.Sprintf("invocationSignature parameter name %q is declared more than once", name),
-			))
-		} else if name != "" {
-			parametersByName[name] = invocationParameterSummary{name: name, valueMode: strings.TrimSpace(parameter.ValueMode)}
-		}
-
-		keys := []struct {
-			value string
-			field string
-		}{}
-		if externalName := strings.TrimSpace(parameter.ExternalName); externalName != "" {
-			keys = append(keys, struct {
-				value string
-				field string
-			}{value: externalName, field: "externalName"})
-		}
-		for aliasIndex, alias := range parameter.Aliases {
-			trimmedAlias := strings.TrimSpace(alias)
-			if trimmedAlias == "" {
-				continue
-			}
-			keys = append(keys, struct {
-				value string
-				field string
-			}{value: trimmedAlias, field: fmt.Sprintf("aliases[%d]", aliasIndex)})
-		}
-		for _, key := range keys {
-			if previousName, exists := namedKeys[key.value]; exists {
-				targets = append(targets, invocationSignatureParameterTarget(
-					CodeInvocationSignatureDuplicateNamedKey,
-					index,
-					key.field,
-					name,
-					fmt.Sprintf("invocationSignature named argument key %q is already used by parameter %q", key.value, previousName),
-				))
-				continue
-			}
-			namedKeys[key.value] = name
-		}
-
-		for bindingIndex, binding := range parameter.Bindings {
-			kind := strings.TrimSpace(binding.Kind)
-			switch kind {
-			case string(factoryapi.FactoryInvocationParameterBindingKindPositional):
-				positionalSlots = append(positionalSlots, binding.Position)
-				if previousName, exists := positionalNames[binding.Position]; exists {
-					targets = append(targets, invocationSignatureParameterBindingTarget(
-						CodeInvocationSignatureInvalidPositionalOrdering,
-						index,
-						bindingIndex,
-						name,
-						fmt.Sprintf("positional slot %d is already bound to parameter %q", binding.Position, previousName),
-					))
-				} else {
-					positionalNames[binding.Position] = name
-				}
-				if strings.TrimSpace(parameter.ValueMode) == string(factoryapi.FactoryInvocationParameterValueModeVariadic) {
-					variadicPositionalNames = append(variadicPositionalNames, name)
-				}
-			case string(factoryapi.FactoryInvocationParameterBindingKindStdin):
-				stdinParameters = append(stdinParameters, name)
-			case string(factoryapi.FactoryInvocationParameterBindingKindNamedRest):
-				namedRestParameters = append(namedRestParameters, name)
-			}
-		}
-	}
-
-	if len(variadicPositionalNames) > 1 {
-		targets = append(targets, invocationSignatureTarget(
-			CodeInvocationSignatureMultipleVariadicPositionals,
-			"parameters",
-			fmt.Sprintf("invocationSignature declares multiple variadic positional parameters (%s)", strings.Join(variadicPositionalNames, ", ")),
-		))
-	}
-
-	if len(positionalSlots) > 0 {
-		slices.Sort(positionalSlots)
-		for index, slot := range positionalSlots {
-			want := index + 1
-			if slot != want {
-				targets = append(targets, invocationSignatureTarget(
-					CodeInvocationSignatureInvalidPositionalOrdering,
-					"parameters",
-					fmt.Sprintf("invocationSignature positional bindings must start at 1 and stay contiguous; found slot %d where slot %d was expected", slot, want),
-				))
-				break
-			}
-		}
-	}
-
-	if len(stdinParameters) > 1 {
-		targets = append(targets, invocationSignatureTarget(
-			CodeInvocationSignatureInvalidStdinRouting,
-			"parameters",
-			fmt.Sprintf("invocationSignature routes stdin to multiple parameters (%s)", strings.Join(stdinParameters, ", ")),
-		))
-	}
-
-	if len(namedRestParameters) > 1 {
-		targets = append(targets, invocationSignatureTarget(
-			CodeInvocationSignatureInvalidNamedRestShape,
-			"parameters",
-			fmt.Sprintf("invocationSignature declares multiple NAMED_REST parameters (%s)", strings.Join(namedRestParameters, ", ")),
-		))
-	}
-	if policy == string(factoryapi.FactoryInvocationUnknownNamedArgumentPolicyCollect) && len(namedRestParameters) != 1 {
-		targets = append(targets, invocationSignatureTarget(
-			CodeInvocationSignatureInvalidNamedRestShape,
-			"unknownNamedArgumentPolicy",
-			"invocationSignature.unknownNamedArgumentPolicy COLLECT requires exactly one parameter with a NAMED_REST binding",
-		))
-	}
-	if policy != string(factoryapi.FactoryInvocationUnknownNamedArgumentPolicyCollect) && len(namedRestParameters) > 0 {
-		targets = append(targets, invocationSignatureTarget(
-			CodeInvocationSignatureInvalidNamedRestShape,
-			"unknownNamedArgumentPolicy",
-			"invocationSignature parameters can only use NAMED_REST when unknownNamedArgumentPolicy is COLLECT",
-		))
-	}
-
-	if signature.OutputContract != nil {
-		targets = append(targets, invocationSignatureOutputContractTargets(signature.OutputContract, parametersByName)...)
-	}
-	targets = append(targets, invocationSignatureInterpolationTargets(cfg, parametersByName)...)
-
-	return targets
-}
-
-func invocationSignatureParameterTargets(parameter interfaces.InvocationParameterConfig, index int) []Target {
-	var targets []Target
-	name := strings.TrimSpace(parameter.Name)
-	valueMode := strings.TrimSpace(parameter.ValueMode)
-	typeHint := strings.TrimSpace(parameter.TypeHint)
-
-	if typeHint != "" && !slices.Contains(supportedInvocationTypeHints, typeHint) {
-		targets = append(targets, invocationSignatureParameterTarget(
-			CodeInvocationSignatureUnsupportedTypeHint,
-			index,
-			"typeHint",
-			name,
-			fmt.Sprintf("invocationSignature parameter %q uses unsupported typeHint %q", name, parameter.TypeHint),
-		))
-	}
-	if valueMode != "" && !slices.Contains(supportedInvocationValueModes, valueMode) {
-		targets = append(targets, invocationSignatureParameterTarget(
-			CodeInvocationSignatureUnsupportedValueMode,
-			index,
-			"valueMode",
-			name,
-			fmt.Sprintf("invocationSignature parameter %q uses unsupported valueMode %q", name, parameter.ValueMode),
-		))
-	}
-	if parameter.DefaultValue != "" && len(parameter.DefaultValues) > 0 {
-		targets = append(targets, invocationSignatureParameterTarget(
-			CodeInvocationSignatureInvalidDefaultShape,
-			index,
-			"defaultValue",
-			name,
-			fmt.Sprintf("invocationSignature parameter %q cannot declare both defaultValue and defaultValues", name),
-		))
-	}
-	switch valueMode {
-	case "", string(factoryapi.FactoryInvocationParameterValueModeExact), string(factoryapi.FactoryInvocationParameterValueModeFileContents):
-		if len(parameter.DefaultValues) > 0 {
-			targets = append(targets, invocationSignatureParameterTarget(
-				CodeInvocationSignatureInvalidDefaultShape,
-				index,
-				"defaultValues",
-				name,
-				fmt.Sprintf("invocationSignature parameter %q only supports defaultValue for its single-value mode", name),
-			))
-		}
-	case string(factoryapi.FactoryInvocationParameterValueModeRepeated), string(factoryapi.FactoryInvocationParameterValueModeVariadic):
-		if parameter.DefaultValue != "" {
-			targets = append(targets, invocationSignatureParameterTarget(
-				CodeInvocationSignatureInvalidDefaultShape,
-				index,
-				"defaultValue",
-				name,
-				fmt.Sprintf("invocationSignature parameter %q only supports defaultValues for its multi-value mode", name),
-			))
-		}
-	}
-	if len(parameter.Choices) > 0 {
-		if parameter.DefaultValue != "" && !slices.Contains(parameter.Choices, parameter.DefaultValue) {
-			targets = append(targets, invocationSignatureParameterTarget(
-				CodeInvocationSignatureInvalidDefaultChoice,
-				index,
-				"defaultValue",
-				name,
-				fmt.Sprintf("invocationSignature parameter %q defaultValue %q is not one of the declared choices", name, parameter.DefaultValue),
-			))
-		}
-		for defaultIndex, defaultValue := range parameter.DefaultValues {
-			if slices.Contains(parameter.Choices, defaultValue) {
-				continue
-			}
-			targets = append(targets, invocationSignatureParameterTarget(
-				CodeInvocationSignatureInvalidDefaultChoice,
-				index,
-				fmt.Sprintf("defaultValues[%d]", defaultIndex),
-				name,
-				fmt.Sprintf("invocationSignature parameter %q defaultValues[%d] %q is not one of the declared choices", name, defaultIndex, defaultValue),
-			))
-		}
-	}
-
-	hasPositionalBinding := false
-	hasNamedRestBinding := false
-	hasNamedBinding := false
-	hasStdinBinding := false
-	for bindingIndex, binding := range parameter.Bindings {
-		kind := strings.TrimSpace(binding.Kind)
-		if kind == "" || !slices.Contains(supportedInvocationBindingKinds, kind) {
-			targets = append(targets, invocationSignatureParameterBindingTarget(
-				CodeInvocationSignatureUnsupportedBindingKind,
-				index,
-				bindingIndex,
-				name,
-				fmt.Sprintf("invocationSignature parameter %q uses unsupported binding kind %q", name, binding.Kind),
-			))
-			continue
-		}
-		switch kind {
-		case string(factoryapi.FactoryInvocationParameterBindingKindPositional):
-			hasPositionalBinding = true
-			if binding.Position < 1 {
-				targets = append(targets, invocationSignatureParameterBindingTarget(
-					CodeInvocationSignatureInvalidPositionalOrdering,
-					index,
-					bindingIndex,
-					name,
-					fmt.Sprintf("invocationSignature parameter %q positional bindings must use positions starting at 1", name),
-				))
-			}
-		case string(factoryapi.FactoryInvocationParameterBindingKindNamed):
-			hasNamedBinding = true
-		case string(factoryapi.FactoryInvocationParameterBindingKindStdin):
-			hasStdinBinding = true
-		case string(factoryapi.FactoryInvocationParameterBindingKindNamedRest):
-			hasNamedRestBinding = true
-		}
-	}
-
-	if parameter.Sensitive && hasPositionalBinding {
-		targets = append(targets, invocationSignatureParameterTarget(
-			CodeInvocationSignatureSensitivePositional,
-			index,
-			"bindings",
-			name,
-			fmt.Sprintf("invocationSignature parameter %q is sensitive and cannot be exposed as a positional argument", name),
-		))
-	}
-	if hasStdinBinding && hasNamedBinding {
-		targets = append(targets, invocationSignatureParameterTarget(
-			CodeInvocationSignatureInvalidStdinRouting,
-			index,
-			"bindings",
-			name,
-			fmt.Sprintf("invocationSignature parameter %q cannot combine STDIN and NAMED bindings", name),
-		))
-	}
-	if hasStdinBinding && hasNamedRestBinding {
-		targets = append(targets, invocationSignatureParameterTarget(
-			CodeInvocationSignatureInvalidStdinRouting,
-			index,
-			"bindings",
-			name,
-			fmt.Sprintf("invocationSignature parameter %q cannot combine STDIN and NAMED_REST bindings", name),
-		))
-	}
-	if hasNamedRestBinding {
-		if valueMode != string(factoryapi.FactoryInvocationParameterValueModeRepeated) {
-			targets = append(targets, invocationSignatureParameterTarget(
-				CodeInvocationSignatureInvalidNamedRestShape,
-				index,
-				"bindings",
-				name,
-				fmt.Sprintf("invocationSignature parameter %q must use valueMode REPEATED when it declares a NAMED_REST binding", name),
-			))
-		}
-		if hasPositionalBinding || hasNamedBinding || hasStdinBinding || len(parameter.Bindings) != 1 {
-			targets = append(targets, invocationSignatureParameterTarget(
-				CodeInvocationSignatureInvalidNamedRestShape,
-				index,
-				"bindings",
-				name,
-				fmt.Sprintf("invocationSignature parameter %q must dedicate its bindings to NAMED_REST only", name),
-			))
-		}
-	}
-	if valueMode == string(factoryapi.FactoryInvocationParameterValueModeRepeated) {
-		if hasPositionalBinding || hasStdinBinding {
-			targets = append(targets, invocationSignatureParameterTarget(
-				CodeInvocationSignatureInvalidRepeatedBindingShape,
-				index,
-				"bindings",
-				name,
-				fmt.Sprintf("invocationSignature parameter %q with valueMode REPEATED may only use NAMED or NAMED_REST bindings", name),
-			))
-		}
-	}
-	if valueMode == string(factoryapi.FactoryInvocationParameterValueModeVariadic) {
-		if !hasPositionalBinding || hasNamedBinding || hasStdinBinding || hasNamedRestBinding {
-			targets = append(targets, invocationSignatureParameterTarget(
-				CodeInvocationSignatureInvalidRepeatedBindingShape,
-				index,
-				"bindings",
-				name,
-				fmt.Sprintf("invocationSignature parameter %q with valueMode VARIADIC must use POSITIONAL bindings only", name),
-			))
-		}
-	}
-
-	return targets
-}
-
-func invocationSignatureOutputContractTargets(
-	output *interfaces.InvocationOutputContractConfig,
-	parameters map[string]invocationParameterSummary,
-) []Target {
-	var targets []Target
-	mode := strings.TrimSpace(output.Mode)
-	if mode != "" && !slices.Contains(supportedOutputContractModes, mode) {
-		targets = append(targets, invocationSignatureTarget(
-			CodeInvocationSignatureUnsupportedOutputContractMode,
-			"outputContract.mode",
-			fmt.Sprintf("invocationSignature.outputContract.mode %q is not supported", output.Mode),
-		))
-	}
-	pathParameter := strings.TrimSpace(output.PathParameter)
-	if pathParameter == "" {
-		return targets
-	}
-	parameter, ok := parameters[pathParameter]
-	if !ok {
-		targets = append(targets, invocationSignatureTarget(
-			CodeInvocationSignatureUnknownOutputPathParameter,
-			"outputContract.pathParameter",
-			fmt.Sprintf("invocationSignature.outputContract.pathParameter %q does not match a declared invocation parameter", pathParameter),
-		))
-		return targets
-	}
-	if !invocationParameterSupportsScalarInterpolation(parameter.valueMode) {
-		targets = append(targets, invocationSignatureTarget(
-			CodeInvocationSignatureInvalidOutputPathParameter,
-			"outputContract.pathParameter",
-			fmt.Sprintf("invocationSignature.outputContract.pathParameter %q must reference a single-value parameter", parameter.name),
-		))
-	}
-	return targets
-}
-
-func invocationSignatureInterpolationTargets(
-	cfg *interfaces.FactoryConfig,
-	parameters map[string]invocationParameterSummary,
-) []Target {
-	fields := invocationInterpolationFieldTargets(cfg)
-	var targets []Target
-	for _, field := range fields {
-		references := invocationSignatureReferencePattern.FindAllStringSubmatch(field.value, -1)
-		for _, reference := range references {
-			parameterName := reference[1]
-			parameter, ok := parameters[parameterName]
-			if !ok {
-				targets = append(targets, Target{
-					Code:     CodeInvocationSignatureInvalidInterpolationReference,
-					Severity: SeverityError,
-					Message:  fmt.Sprintf("%s references invocation parameter %q, but that parameter is not declared", field.fieldDescriptor, parameterName),
-					Subject: Subject{
-						Type:     field.subjectType,
-						ID:       field.subjectID,
-						Location: field.location,
-					},
-					Path: field.path,
-				})
-				continue
-			}
-			if field.allowsRepeated || invocationParameterSupportsScalarInterpolation(parameter.valueMode) {
-				continue
-			}
-			targets = append(targets, Target{
-				Code:     CodeInvocationSignatureIncompatibleInterpolationReference,
-				Severity: SeverityError,
-				Message:  fmt.Sprintf("%s cannot reference multi-value invocation parameter %q", field.fieldDescriptor, parameterName),
-				Subject: Subject{
-					Type:     field.subjectType,
-					ID:       field.subjectID,
-					Location: field.location,
-				},
-				Path: field.path,
-			})
-		}
-	}
-	return targets
-}
-
-func invocationInterpolationFieldTargets(cfg *interfaces.FactoryConfig) []interpolationFieldTarget {
-	var fields []interpolationFieldTarget
-	for workerIndex, worker := range cfg.Workers {
-		subjectID := worker.Name
-		basePath := fmt.Sprintf("%s.workers[%d](%s)", validationRoot, workerIndex, worker.Name)
-		fields = appendInterpolationField(fields, SubjectTypeWorker, subjectID, SubjectLocationDefinition, basePath+".provider", worker.Provider, false, "worker.provider")
-		fields = appendInterpolationField(fields, SubjectTypeWorker, subjectID, SubjectLocationDefinition, basePath+".model", worker.Model, false, "worker.model")
-		fields = appendInterpolationField(fields, SubjectTypeWorker, subjectID, SubjectLocationDefinition, basePath+".modelProvider", worker.ModelProvider, false, "worker.modelProvider")
-		fields = appendInterpolationField(fields, SubjectTypeWorker, subjectID, SubjectLocationDefinition, basePath+".executorProvider", worker.ExecutorProvider, false, "worker.executorProvider")
-		fields = appendInterpolationField(fields, SubjectTypeWorker, subjectID, SubjectLocationDefinition, basePath+".command", worker.Command, false, "worker.command")
-		fields = appendInterpolationField(fields, SubjectTypeWorker, subjectID, SubjectLocationDefinition, basePath+".timeout", worker.Timeout, false, "worker.timeout")
-		fields = appendInterpolationField(fields, SubjectTypeWorker, subjectID, SubjectLocationDefinition, basePath+".stopToken", worker.StopToken, false, "worker.stopToken")
-		fields = appendInterpolationField(fields, SubjectTypeWorker, subjectID, SubjectLocationDefinition, basePath+".openCodeAgent", worker.OpenCodeAgent, false, "worker.openCodeAgent")
-		fields = appendInterpolationField(fields, SubjectTypeWorker, subjectID, SubjectLocationDefinition, basePath+".body", worker.Body, false, "worker body")
-		for argIndex, arg := range worker.Args {
-			fields = appendInterpolationField(fields, SubjectTypeWorker, subjectID, SubjectLocationDefinition, fmt.Sprintf("%s.args[%d]", basePath, argIndex), arg, false, "worker.args entry")
-		}
-	}
-	for workstationIndex, workstation := range cfg.Workstations {
-		subjectID := workstation.Name
-		basePath := fmt.Sprintf("%s.workstations[%d](%s)", validationRoot, workstationIndex, workstation.Name)
-		fields = appendInterpolationField(fields, SubjectTypeWorkstation, subjectID, SubjectLocationDefinition, basePath+".worker", workstation.WorkerTypeName, false, "workstation.worker")
-		fields = appendInterpolationField(fields, SubjectTypeWorkstation, subjectID, SubjectLocationDefinition, basePath+".runner", workstation.Runner, false, "workstation.runner")
-		fields = appendInterpolationField(fields, SubjectTypeWorkstation, subjectID, SubjectLocationDefinition, basePath+".openCodeAgent", workstation.OpenCodeAgent, false, "workstation.openCodeAgent")
-		fields = appendInterpolationField(fields, SubjectTypeWorkstation, subjectID, SubjectLocationDefinition, basePath+".promptFile", workstation.PromptFile, false, "workstation.promptFile")
-		fields = appendInterpolationField(fields, SubjectTypeWorkstation, subjectID, SubjectLocationDefinition, basePath+".outputSchema", workstation.OutputSchema, false, "workstation.outputSchema")
-		fields = appendInterpolationField(fields, SubjectTypeWorkstation, subjectID, SubjectLocationDefinition, basePath+".timeout", workstation.Timeout, false, "workstation.timeout")
-		fields = appendInterpolationField(fields, SubjectTypeWorkstation, subjectID, SubjectLocationDefinition, basePath+".body", workstation.Body, false, "workstation prompt body")
-		fields = appendInterpolationField(fields, SubjectTypeWorkstation, subjectID, SubjectLocationDefinition, basePath+".promptTemplate", workstation.PromptTemplate, false, "workstation.promptTemplate")
-		fields = appendInterpolationField(fields, SubjectTypeWorkstation, subjectID, SubjectLocationDefinition, basePath+".workingDirectory", workstation.WorkingDirectory, false, "workstation.workingDirectory")
-		fields = appendInterpolationField(fields, SubjectTypeWorkstation, subjectID, SubjectLocationDefinition, basePath+".worktree", workstation.Worktree, false, "workstation.worktree")
-		for key, value := range workstation.Env {
-			fields = appendInterpolationField(fields, SubjectTypeWorkstation, subjectID, SubjectLocationDefinition, fmt.Sprintf("%s.env[%q]", basePath, key), value, false, fmt.Sprintf("workstation.env[%q]", key))
-		}
-	}
-	return fields
-}
-
-func appendInterpolationField(
-	fields []interpolationFieldTarget,
-	subjectType SubjectType,
-	subjectID string,
-	location SubjectLocation,
-	path string,
-	value string,
-	allowsRepeated bool,
-	fieldDescriptor string,
-) []interpolationFieldTarget {
-	if !strings.Contains(value, "${") {
-		return fields
-	}
-	return append(fields, interpolationFieldTarget{
-		subjectType:     subjectType,
-		subjectID:       subjectID,
-		path:            path,
-		location:        location,
-		value:           value,
-		allowsRepeated:  allowsRepeated,
-		fieldDescriptor: fieldDescriptor,
-	})
-}
-
-func invocationParameterSupportsScalarInterpolation(valueMode string) bool {
-	switch valueMode {
-	case string(factoryapi.FactoryInvocationParameterValueModeRepeated), string(factoryapi.FactoryInvocationParameterValueModeVariadic):
-		return false
-	default:
-		return true
-	}
-}
-
-func invocationSignatureTarget(code, field, message string) Target {
-	return Target{
-		Code:     code,
-		Severity: SeverityError,
-		Message:  message,
-		Subject: Subject{
-			Type:     SubjectTypeFactory,
-			ID:       "invocationSignature",
-			Location: SubjectLocationDefinition,
-		},
-		Path: fmt.Sprintf("%s.invocationSignature.%s", validationRoot, field),
-	}
-}
-
-func invocationSignatureParameterTarget(code string, index int, field string, name string, message string) Target {
-	subjectID := name
-	if subjectID == "" {
-		subjectID = fmt.Sprintf("parameter[%d]", index)
-	}
-	return Target{
-		Code:     code,
-		Severity: SeverityError,
-		Message:  message,
-		Subject: Subject{
-			Type:     SubjectTypeFactory,
-			ID:       subjectID,
-			Location: SubjectLocationDefinition,
-		},
-		Path: fmt.Sprintf("%s.invocationSignature.parameters[%d].%s", validationRoot, index, field),
-	}
-}
-
-func invocationSignatureParameterBindingTarget(code string, parameterIndex int, bindingIndex int, name string, message string) Target {
-	subjectID := name
-	if subjectID == "" {
-		subjectID = fmt.Sprintf("parameter[%d]", parameterIndex)
-	}
-	return Target{
-		Code:     code,
-		Severity: SeverityError,
-		Message:  message,
-		Subject: Subject{
-			Type:     SubjectTypeFactory,
-			ID:       subjectID,
-			Location: SubjectLocationDefinition,
-		},
-		Path: fmt.Sprintf("%s.invocationSignature.parameters[%d].bindings[%d]", validationRoot, parameterIndex, bindingIndex),
 	}
 }
