+++ ui/src/features/dashboard/lib/dashboard-session-lifecycle.test.ts
@@ -62,8 +62,9 @@ describe("shouldResetDashboardSessionScopedState", () => {
 });
 
 describe("resetDashboardSessionScopedState", () => {
-  it("resets timeline, stream, selection history, and factory-definition queries once", () => {
+  it("removes factory-definition queries when requested", () => {
     const queryClient = {
+      invalidateQueries: vi.fn(),
       removeQueries: vi.fn(),
     };
     const resetTimeline = vi.fn();
@@ -80,14 +81,40 @@ describe("resetDashboardSessionScopedState", () => {
     expect(resetStreamState).toHaveBeenCalledTimes(1);
     expect(resetStreamState).toHaveBeenCalledWith("en");
     expect(queryClient.removeQueries).toHaveBeenCalledTimes(1);
+    expect(queryClient.invalidateQueries).not.toHaveBeenCalled();
     expect(queryClient.removeQueries).toHaveBeenCalledWith({
       queryKey: [CURRENT_FACTORY_DEFINITION_QUERY_KEY_PREFIX],
       exact: false,
     });
   });
 
+  it("invalidates factory-definition queries without dropping cached data when requested", () => {
+    const queryClient = {
+      invalidateQueries: vi.fn(),
+      removeQueries: vi.fn(),
+    };
+    const resetTimeline = vi.fn();
+    const resetStreamState = vi.fn();
+
+    resetDashboardSessionScopedState(
+      queryClient as never,
+      resetStreamState,
+      resetTimeline,
+      "en",
+      "invalidate",
+    );
+
+    expect(queryClient.invalidateQueries).toHaveBeenCalledTimes(1);
+    expect(queryClient.invalidateQueries).toHaveBeenCalledWith({
+      queryKey: [CURRENT_FACTORY_DEFINITION_QUERY_KEY_PREFIX],
+      exact: false,
+    });
+    expect(queryClient.removeQueries).not.toHaveBeenCalled();
+  });
+
   it("resets stream state without locale when omitted", () => {
     const queryClient = {
+      invalidateQueries: vi.fn(),
       removeQueries: vi.fn(),
     };
     const resetTimeline = vi.fn();
