+++ ui/src/features/dashboard/lib/dashboard-session-lifecycle.ts
@@ -8,6 +8,8 @@ import {
 import { resetSelectionHistoryStore } from "../../current-selection/base/public";
 import { FACTORY_SESSION_DETAIL_QUERY_KEY } from "../../factory-session-detail/public";
 
+export type FactoryDefinitionQueryResetMode = "invalidate" | "remove";
+
 export function dashboardSessionKey(
   sessionID: string | null,
   refreshToken: number,
@@ -35,14 +37,20 @@ export function resetDashboardSessionScopedState(
   resetStreamState: (locale?: string | null) => void,
   resetTimeline: () => void,
   locale?: string | null,
+  factoryDefinitionQueryResetMode: FactoryDefinitionQueryResetMode = "remove",
 ): void {
   resetTimeline();
   resetStreamState(locale);
   resetSelectionHistoryStore();
-  queryClient.removeQueries({
+  const queryFilter = {
     queryKey: [CURRENT_FACTORY_DEFINITION_QUERY_KEY_PREFIX],
     exact: false,
-  });
+  } as const;
+  if (factoryDefinitionQueryResetMode === "invalidate") {
+    void queryClient.invalidateQueries(queryFilter);
+    return;
+  }
+  queryClient.removeQueries(queryFilter);
 }
 
 export function clearDashboardSessionRuntimeQueries(
